New JSON harakiri protocol
This defines a new simple JSON protocol for harakiri. This allows
commands to be sent as JSON and to receive appropriate responses. The
protocol is not "streaming" per se as JSON objects are still delimited
via newlines, however it should be simple enough to extract the
responses.

See http://www.couchbase.com/wiki/display/couchbase/CouchbaseMock
and subpages for more information on the specs of this protocol.

Change-Id: Ibbfa2b356804f0b40ff8b1b2ad89c05b04bbe903
Reviewed-on: http://review.couchbase.org/27120
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Trond Norbye <trond.norbye@gmail.com>