IO Refactor for memcached
This fixes some issues in the Memcached I/O.
Previously we were using a simple write with nonblocking I/O; where we'd
issue a chan.write() call and discard the buffer regardless of whether
all bytes were written or not. This caused data to get lost if the TCP
write buffer was full and caused me lots of frustration on Windows
(where this happened quite frequently).

Change-Id: Ib7ae9b3a82f6818697032b1a17f33a98f4003bbd
Reviewed-on: http://review.couchbase.org/28120
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: Mark Nunberg <mnunberg@haskalah.org>