Add Couchbase Views
This draws upon some work of Sergey Avseyev (who provided the initial
view implementation) and Brett Lawson (who provided the core of the
bundled javascript implementation).

This utilizes a mostly-in-javascript implementation of indexing and
returning documents per the view interface.

It adds a hard dependency on the Rhino Javascript engine (skipping
entirely over any ScriptEngine implementation built in to JDKs, as they
are stripped down and unreliable for our purpose).

Change-Id: Ifee524e750f034111dad467211419c93711eb5f6
Reviewed-on: http://review.couchbase.org/44726
Tested-by: Mark Nunberg <mnunberg@haskalah.org>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>