Add support for "Fixed" n1ql queries
This adds support for basic n1ql "Execution" and "Preparation" and
"Index States". This actually works by understanding a single statement,
"SELECT mockrow" which can also be prepared.
To emulate "index states", the "prepared" statement contains a random
number which reflects a "known index". The number can be changed
out-of-band to simulate a change in index states.

Change-Id: I5976e4f70043dfcdf24f8e11761b522547c434d7
Reviewed-on: http://review.couchbase.org/54791
Reviewed-by: Subhashni Balakrishnan <b.subhashni@gmail.com>
Tested-by: Mark Nunberg <mark.nunberg@couchbase.com>