Compile for "-release 8"; Require Java 9 to build
Motivation
----------
I botched the 1.5.26 release because I built it with
Java 17. Even though the target and source properties
were correctly set to "1.8" and the generated class
files were compatible with Java 1.8, the code failed
at runtime on a 1.8 JVM because of changes to
how ByteBuffer was implemented between Java 8 and 9.

The best fix for this issue is to specify the
"-release" option when compiling the code. This option
ensures binary compatibility with the specified
version, regardless of the version of the JDK used to
compile the code.

The downside is that this option is only supported
by JDK 9 and later. This means Java 9 or later is now
required to build the mock. The mock remains compatible
with Java 8 at runtime.

Modifications
-------------
When compiling, specify the "release" option instead
of source and target language levels.

Bump Maven plugin versions.

Change-Id: Ie40be7f94977cd159433faaf8b7b663bc40dc6fb
Reviewed-on: https://review.couchbase.org/c/CouchbaseMock/+/193604
Reviewed-by: Michael Reiche <michael.reiche@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Tested-by: David Nault <david.nault@couchbase.com>